Fix manage token scene layout issues
Use wrapper view around the buttons to enforce proper component spacing.
Remove gap / space below the scene header.
Remove unnecessary view components.
Replace the `SceneFooter` with a `DividerLine` component.
